Создали математическую модель продуктового магазина

+++ src/myclasses/Product.java
@@ -25,7 +25,7 @@ public class Product {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
